Improved the priority system
- Domain module does not depend on compose (note hardcoded color!)
- Removed dangerous call when opening the color picker
