+++ core/model/src/main/java/com/axiel7/anihyou/core/model/media/PriorityFormat.kt
@@ -3,6 +3,7 @@ package com.axiel7.anihyou.core.model.media
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
+import com.axiel7.anihyou.core.resources.R
 import com.axiel7.anihyou.core.resources.md_theme_dark_outline
 import com.axiel7.anihyou.core.resources.md_theme_light_outline
 import com.axiel7.anihyou.core.resources.stat_dark_10
@@ -22,4 +23,12 @@ fun Int.priorityColor(): Color {
         2 -> if (isDark) stat_dark_50 else stat_light_50
         else -> if (isDark) md_theme_dark_outline else md_theme_light_outline
     }
+}
+
+@Composable
+fun Int.priorityIcon(): Int = when (this) {
+    0 -> R.drawable.counter_0_24
+    1 -> R.drawable.counter_1_24
+    2 -> R.drawable.counter_2_24
+    else -> R.drawable.cancel_24 // invalid priority was set
 }
