+++ core/ui/src/main/java/com/axiel7/anihyou/core/ui/composables/scores/MediaListScoreIndicator.kt
@@ -16,7 +16,6 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalLocale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -31,7 +30,6 @@ import com.axiel7.anihyou.core.model.smileyOnPrimaryColor
 import com.axiel7.anihyou.core.model.smileyPrimaryColor
 import com.axiel7.anihyou.core.network.type.ScoreFormat
 import com.axiel7.anihyou.core.resources.R
-import com.materialkolor.ktx.harmonize
 
 @Composable
 fun BadgeScoreIndicator(
@@ -176,45 +174,5 @@ fun MinimalScoreIndicator(
 
             ScoreFormat.UNKNOWN__ -> {}
         }
-    }
-}
-
-@Composable
-fun PriorityIndicator(
-    modifier: Modifier,
-    priority: Int,
-    lowPriorityColor: Color? = null,
-    mediumPriorityColor: Color? = null,
-    highPriorityColor: Color? = null,
-) {
-    val shape = RoundedCornerShape(topEnd = 8.dp, bottomStart = 16.dp)
-    val iconId = when (priority) {
-        0 -> R.drawable.counter_0_24
-        1 -> R.drawable.counter_1_24
-        2 -> R.drawable.counter_2_24
-        else -> R.drawable.cancel_24 // invalid priority was set
-    }
-    val backgroundColor = when (priority) {
-        0 -> lowPriorityColor ?: MaterialTheme.colorScheme.secondaryContainer
-        1 -> mediumPriorityColor ?: MaterialTheme.colorScheme.secondaryContainer
-        2 -> highPriorityColor ?: MaterialTheme.colorScheme.secondaryContainer
-        else -> MaterialTheme.colorScheme.secondaryContainer
-    }
-    val icon = painterResource(iconId)
-
-
-    Row(
-        modifier = modifier
-            .clip(shape)
-            .background(backgroundColor.harmonize(MaterialTheme.colorScheme.primary))
-            .padding(horizontal = 8.dp, vertical = 4.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            painter = icon,
-            contentDescription = stringResource(R.string.priority),
-            modifier = Modifier.size(20.dp),
-            tint = MaterialTheme.colorScheme.secondaryContainer
-        )
     }
 }
