+++ gradle/libs.versions.toml
@@ -53,7 +53,6 @@ haze = "2.0.0-alpha03"
 reorderable = "3.1.0"
 zoomable = "2.13.0"
 colorpicker = "1.2.0"
-material3Version = "1.4.0"
 
 [libraries]
 accompanist-permissions = { group = "com.google.accompanist", name = "accompanist-permissions", version.ref = "accompanist" }
@@ -127,7 +126,6 @@ haze-blur = { group = "dev.chrisbanes.haze", name = "haze-blur", version.ref = "
 reorderable = { module = "sh.calvin.reorderable:reorderable", version.ref = "reorderable" }
 zoomable = { module = "net.engawapg.lib:zoomable", version.ref = "zoomable" }
 compose-colorpicker = { module = "com.github.skydoves:colorpicker-compose", version.ref = "colorpicker" }
-androidx-compose-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "material3Version" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
